blockchain-client/requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyasn1==0.4.8
pycryptodomex==3.14.1
rsa==4.8
Werkzeug==2.1.2
zipp==3.8.0
